requirements.txt
opencv-python==4.9.0.80
pyTelegramBotAPI==4.16.1
python-dotenv==1.0.1
pyzbar==0.1.9
